Split domain helpers out of ApplicationHelper
Rails mixes every helper into every view, so helper names share one global
namespace; ApplicationHelper had become the default dumping ground for methods
that serve one domain. It now holds page chrome only, down from 278 to 100
lines.

- New modules: IconsHelper (nav_icon, action_icon), NavigationHelper,
  CurrencyHelper, CountriesHelper, DeliveryNotesHelper.
- ActionButtonsHelper absorbs the generic building blocks its per-verb wrappers
  were already built from, and declares its IconsHelper dependency.
- status_badge_tag serves delivery notes alone and is renamed
  delivery_note_status_badge_tag to match.
- EmailPreviewHelper splits by layer: mail extraction is controller-side and
  moves to the EmailPreviewExtraction concern, so it no longer leaks into every
  view; the helper keeps the view-side Stimulus wiring.
- icon_label is public — seven views already call it.
- dropdown_link_to is deleted; it had no call sites.

Helper tests follow their methods into matching test files.
