+++ app/helpers/email_preview_helper.rb
@@ -1,44 +1,16 @@
 module EmailPreviewHelper
-  # Extracts email preview data from ActionMailer::MessageDelivery or Mail::Message
-  def extract_email_preview_data(mail)
-    data = {
-      to: mail.to&.join(", "),
-      from: mail.from&.first,
-      subject: mail.subject,
-      text_body: extract_text_body(mail),
-      attachments: extract_attachments_info(mail)
+  # Stimulus data attributes that wire an element up to the
+  # generic-email-preview controller for a given Invoice or DeliveryNote.
+  # Relies on the routes following the {action}_{resource}_path convention.
+  # The controller-side counterpart — building the previewed payload — is
+  # EmailPreviewExtraction.
+  def email_preview_data(resource)
+    prefix = resource.class.name.underscore
+    {
+      controller: "generic-email-preview",
+      "generic-email-preview-preview-url-value" => send("preview_email_#{prefix}_path", resource),
+      "generic-email-preview-html-preview-url-value" => send("preview_email_html_#{prefix}_path", resource),
+      "generic-email-preview-send-url-value" => send("send_email_#{prefix}_path", resource)
     }
-
-    data[:cc] = mail.cc&.join(", ") if mail.cc&.any?
-    data[:bcc] = mail.bcc&.join(", ") if mail.bcc&.any?
-    data
-  end
-
-  private
-
-  def extract_html_body(mail)
-    if mail.multipart?
-      mail.html_part&.body&.decoded
-    else
-      mail.content_type&.include?("text/html") ? mail.body.decoded : nil
-    end
-  end
-
-  def extract_text_body(mail)
-    if mail.multipart?
-      mail.text_part&.body&.decoded
-    else
-      mail.content_type&.include?("text/plain") ? mail.body.decoded : nil
-    end
-  end
-
-  def extract_attachments_info(mail)
-    mail.attachments.map do |attachment|
-      {
-        filename: attachment.filename,
-        content_type: attachment.content_type,
-        size: attachment.body.raw_source.bytesize
-      }
-    end
   end
 end
