+++ app/views/delivery_notes/index.html.haml
@@ -89,7 +89,7 @@
           %td= l(delivery_note.delivery_end_date || delivery_note.delivery_start_date)
           %td= delivery_note.internal_reference
           %td= delivery_note.cust_order
-          %td= status_badge_tag(delivery_note)
+          %td= delivery_note_status_badge_tag(delivery_note)
           %td
             - if !delivery_note.published? && can?('delivery_notes.edit')
               = list_action_link('Edit', edit_delivery_note_path(delivery_note), :edit)
