+++ docs/code-style.md
@@ -55,7 +55,7 @@ Three tiers:
 - **Icon + text** for less-familiar workflow actions (`Publish`, `Upload PDF`, `Block`).
 - **Icon only** for universally-understood verbs. Always pass `title:` (sets tooltip + `aria-label`): Delete, PDF, Preview.
 
-Icons are inline monochrome SVG via `action_icon(:name)` (`app/helpers/application_helper.rb`), backed by the `bootstrap-icons` gem — not emoji. `action_icon` takes the gem's icon name (underscores are converted to hyphens, so `:"eye-fill"` or `:eye_fill` both map to the `eye-fill` icon). Build an icon + text label with `icon_label(:name, "Text")` (`app/helpers/action_buttons_helper.rb`) — it wraps the text in a `<span>` so `.btn svg.bi:not(:last-child)` in `bootstrap_and_overrides.css.scss` only adds the icon/text gap where there's text to gap against, keeping icon-only buttons symmetrically centered. `<input type="submit">` can't hold an SVG — use `f.button(...) do ... end` instead of `f.submit` wherever the label needs an icon.
+Icons are inline monochrome SVG via `action_icon(:name)` (`app/helpers/icons_helper.rb`), backed by the `bootstrap-icons` gem — not emoji. `action_icon` takes the gem's icon name (underscores are converted to hyphens, so `:"eye-fill"` or `:eye_fill` both map to the `eye-fill` icon). Build an icon + text label with `icon_label(:name, "Text")` (`app/helpers/action_buttons_helper.rb`) — it wraps the text in a `<span>` so `.btn svg.bi:not(:last-child)` in `bootstrap_and_overrides.css.scss` only adds the icon/text gap where there's text to gap against, keeping icon-only buttons symmetrically centered. `<input type="submit">` can't hold an SVG — use `f.button(...) do ... end` instead of `f.submit` wherever the label needs an icon.
 
 When the icon carries the verb, drop the leading verb from the label: `Mark Paid` → `Paid…`, `Mark Unpaid` → `Unpaid`. Keep verb+object when the icon alone isn't enough.
